+++ src/features/CardPost/index.ts
@@ -1,0 +1,1 @@
+export {CardPost} from '@features/CardPost/ui/CardPost'
