+++ src/features/CardPost/ui/CardPost.tsx
@@ -1,45 +1,35 @@
 import { CommentOutlined, HeartFilled, HeartOutlined } from '@ant-design/icons';
-import { postSlice } from '@features/posts/module/reduce';
-import { useAppDispatch } from '@providers/store/hooks';
-import { CardType } from '@shared/lib/types';
-import style from '@shared/styles/ui/card.module.css';
 import { useState } from 'react';
-import { ModalPost } from './ModalPost';
+import { ModalPost } from 'src/features/ModalPost/ui/ModalPost';
+import { useCardPost } from '../hooks/useCardPost';
+import { CardType } from '../lib/type';
+import style from '../style/card.module.css';
 export const CardPost = ({
 	userImg,
 	userName,
 	data,
 	content,
 	postImg,
-	postLike,
 	postCommentsCount,
 	isLiked,
+	postLike,
 	postId,
 	getLike,
 	deleteLike,
 	getComment,
 }: CardType) => {
-	const dispatch = useAppDispatch();
-	const [isLike, setIsLike] = useState<boolean>(isLiked || false);
-	const [count, setCount] = useState<number>(postLike);
 	const [modalOpen, setModalOpen] = useState<boolean>(false);
-
-	const handleLike = async () => {
-		setIsLike(true);
-		setCount(prev => prev + 1);
-		dispatch(postSlice.actions.targgetLike({ isLike, postId }));
-		await getLike();
-	};
-	const handleUnLike = async () => {
-		setIsLike(false);
-		setCount(prev => prev - 1);
-		dispatch(postSlice.actions.targgetLike({ isLike, postId }));
-		await deleteLike();
-	};
-	const onClickComments = () => {
-		getComment();
-		setModalOpen(true);
-	};
+	const { isLike, count, handleLike, handleUnLike, onClickComments } =
+		useCardPost({
+			isLiked,
+			postLike,
+			getLike,
+			deleteLike,
+			getComment,
+			postId,
+			count: postLike,
+			setModalOpen,
+		});
 	return (
 		<>
 			<div className={style.container}>
@@ -65,7 +55,10 @@ export const CardPost = ({
 						<span className={style.span}>{count}</span>
 					</div>
 					<div className={style.post_comments}>
-						<CommentOutlined className={style.comments} onClick={onClickComments} />
+						<CommentOutlined
+							className={style.comments}
+							onClick={onClickComments}
+						/>
 						<span className={style.span}>{postCommentsCount}</span>
 					</div>
 				</div>
