+++ src/features/ModalPost/lib/type.ts
@@ -1,0 +1,4 @@
+export interface IUseModalPost {
+	postId?:number,
+	setModuleOpen: (open:boolean) => void,
+}
