+++ src/features/posts/index.ts
@@ -1,0 +1,2 @@
+export {Post} from "@features/posts/ui/Post"
+export {getLike,deleteLike,postComments,getComments} from "@features/posts/api/request"
