+++ src/shared/ui/Link.tsx
@@ -1,7 +1,7 @@
-import { LinkType } from '@shared/lib/types';
 import { FC } from 'react';
+import { LinkType } from 'src/shared/lib/types';
 import style from '../styles/ui/links.module.css';
-export const Link: FC<LinkType> = ({ img, text, width, onClick,isActive }) => {
+export const Link: FC<LinkType> = ({ img, text, width, onClick, isActive }) => {
 	return (
 		<div className={style.container}>
 			<div
